requirement.txt
asgiref==3.4.1
asn1crypto==0.24.0
boto3==1.23.10
botocore==1.26.10
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
Django==3.2.23
django-debug-toolbar==3.2.4
django-filter==2.4.0
django-model-utils==4.2.0
django-notifications-hq==1.8.3
django-storages==1.12.3
djangorestframework==3.12.2
idna==2.6
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
mysqlclient==2.0.3
pycrypto==2.6.1
PyGObject==3.26.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyxdg==0.25
requests==2.18.4
s3transfer==0.5.2
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.4.4
ssh-import-id==5.7
swapper==1.3.0
typing_extensions==4.1.1
urllib3==1.26.18
wrapt==1.15.0
